assignments/assignment_4/pyproject.toml
[project]
name = "bertopic-news-analysis"
version = "0.1.0"
description = "Topic analysis of news headlines with BERTopic"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["bertopic>=0.15.0", 
"jsonlines>=3.1.0",
"matplotlib>=3.5.0",
"numpy>=1.20.0",
"pandas>=1.3.0",
"pydantic>=2.0.0",
"pydantic-settings>=2.0.0",
"scikit-learn>=1.0.0",
"seaborn>=0.11.0",
"tqdm>=4.62.0",
"umap-learn>=0.5.0",
"hdbscan>=0.8.29"
]
